resin-proxy-config: Support for dns over TCP and in containers connectivity
Before multicontainer support in resin, the user container was running over the
host's network stack which means that all the packets were treated as locally
emitted ones. In this case, the communication was redirected as per the OUTPUT
iptable rule only.
In the "multicontainer world" the user containers run on top of a bridge network
which makes the host act as a socksifying router. In this case, the packets
won't go through the OUTPUT chain and the redirection to redsocks needs to
happen at PREROUTING chain.
This change adds a redirect rule at PREROUTING for TCP packets to redsocks to
redirect packets from the containers to redsocks too.

As well, this adds support for redirecting DNS UDP packets to redsocks. This is
implemented as iptables rules which are only added if the redsocks
configuration has a `dnsu2t` section.

Change-type: patch
Changelog-entry: Fix proxy when using containers over bridge network
Signed-off-by: Andrei Gherzan <andrei@resin.io>