Removed requirement on docker daemon being a local process
The test that docker is available first checks for a running process, then checks permissions by running a docker command. The first part is unnecessary (because the second part is sufficient) and fails even if there is a usable non-local docker daemon available.
Change-type: patch
Chengelog-entry: Removed requirement on docker daemon being a local process
Signed-off-by: Mark Grimes <mark.grimes@rymapt.com>